requirements.txt
fastapi
uvicorn
jinja2
pymongo
feedparser
pyshorteners
requests
python-dotenv
pyngrok